junit search string

+++ src/lab02/testStrings.java
@@ -6,8 +6,10 @@ public class testStrings {
         System.out.println("abc".substring(1,3));
         //boolean b;
         boolean b=false;
+
+
         System.out.println(b);
-        float f = 2.3f;
+        double f = 2.3;
         //System.out.println(s.substring(7));
         while (b){
             System.out.println("never reached");
